requirements.txt
elevenlabs==1.3.0
python-dotenv==1.0.0
pydub==0.25.1
pygame==2.5.2
requests==2.31.0
rich==13.7.0
sounddevice==0.4.6
numpy==1.26.2
tqdm==4.66.1